Almost done
yeah

+++ book-collection/app/Http/Controllers/BookController.php
@@ -7,6 +7,8 @@ use App\Models\Book;
 use App\Http\Requests\StoreBookRequest;
 use App\Models\Author;
 use App\Models\Review;
+use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Exceptions\HttpResponseException;
 
 class BookController extends Controller
 {
@@ -41,10 +43,26 @@ public function update(StoreBookRequest $request, Book $book) {
     return BookResource::collection($books);
 }
 
-public function destroy(Book $book) {
+// public function destroy(Book $book) {
+//     $book->delete();
+//     return response()->json(['message' => 'Boek succesvol verwijderd']);
+// }
+
+
+
+    public function destroy(Book $book) {
+    if ($book->reviews()->exists()) {
+        throw new HttpResponseException(response()->json([
+            'errors' => [],
+            'message' => 'Dit boek kan niet worden verwijderd omdat er nog reviews aan gekoppeld zijn.'
+        ], 422));
+    }
+
     $book->delete();
-    return response()->json(['message' => 'Boek succesvol verwijderd']);
-}
+    }
+
+
+
 
     public function show() {
